Refactor: Simplify user facing API
